requirements.txt
django
djangorestframework
django-cors-headers
tensorflow
whitenoise
pillow